refactor: rename ui files to pascal and camel case (#26)
