+++ src/components/ui/Field.tsx
@@ -2,8 +2,8 @@ import { useMemo } from 'react'
 import { cva, type VariantProps } from 'class-variance-authority'
 
 import { cn } from '@/lib/utils/cn'
-import { Label } from '@/components/ui/label'
-import { Separator } from '@/components/ui/separator'
+import { Label } from '@/components/ui/Label'
+import { Separator } from '@/components/ui/Separator'
 
 function FieldSet({ className, ...props }: React.ComponentProps<'fieldset'>) {
   return (
