+++ src/components/ui/Item.tsx
@@ -4,7 +4,7 @@ import { useRender } from '@base-ui/react/use-render'
 import { cva, type VariantProps } from 'class-variance-authority'
 
 import { cn } from '@/lib/utils/cn'
-import { Separator } from '@/components/ui/separator'
+import { Separator } from '@/components/ui/Separator'
 
 function ItemGroup({ className, ...props }: React.ComponentProps<'div'>) {
   return (
